Finally moving the SerdeSe artifacts into the serde mod
