+++ components/provider_cldr/src/transform/mod.rs
@@ -11,10 +11,10 @@ pub use plurals::PluralsProvider;
 
 use crate::support::LazyCldrProvider;
 use crate::CldrPaths;
-use icu_provider::erased::SerdeSeDataStruct;
 use icu_provider::erased::*;
 use icu_provider::iter::{IterableDataProviderCore, KeyedDataProvider};
 use icu_provider::prelude::*;
+use icu_provider::serde::SerdeSeDataStruct;
 
 /// Returns a list of all ResourceKeys that this provider can produce.
 pub fn get_all_resc_keys() -> Vec<ResourceKey> {
@@ -45,7 +45,7 @@ impl<'a, 'd> CldrJsonDataProvider<'a, 'd> {
 }
 
 impl<'a, 'd> ErasedDataProvider<'d> for CldrJsonDataProvider<'a, 'd> {
-    fn load_payload(
+    fn load_erased(
         &self,
         req: &DataRequest,
     ) -> Result<DataResponse<'d, dyn ErasedDataStruct>, DataError> {
