+++ components/provider_fs/src/export/fs_exporter.rs
@@ -8,9 +8,9 @@ use crate::manifest::AliasOption;
 use crate::manifest::LocalesOption;
 use crate::manifest::Manifest;
 use crate::manifest::MANIFEST_FILE;
-use icu_provider::erased::SerdeSeDataStruct;
 use icu_provider::export::DataExporter;
 use icu_provider::prelude::*;
+use icu_provider::serde::SerdeSeDataStruct;
 use serde::{Deserialize, Serialize};
 use std::fs;
 use std::io::Write;
